+++ src/overrides/js.js
@@ -1,0 +1,3 @@
+export const jsOverridesRules = {
+  '@typescript-eslint/no-var-requires': 'off',
+}